feat(apollo-wind): add Chat Experience page, rename Ideas to Layout Generator
- Adds Experiments/Chat Experience page (coming soon placeholder)
- Renames Experiments/Ideas to Experiments/Layout Generator
- Removes Experiments/Brand Motion page
