--- packages/apollo-wind/src/templates/Experiments/brand-motion.stories.tsx
@@ -1,56 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react-vite';
-import { fontFamily } from '@/foundation/Future/typography';
-import { cn } from '@/lib';
-
-// ============================================================================
-// Meta
-// ============================================================================
-
-const meta = {
-  title: 'Experiments/Brand Motion',
-  parameters: {
-    layout: 'fullscreen',
-  },
-} satisfies Meta;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-function resolveThemeClass(value: string) {
-  return value ?? 'future-dark';
-}
-
-// ============================================================================
-// Coming Soon placeholder
-// ============================================================================
-
-function ComingSoon({ theme }: { theme: string }) {
-  const themeClass = resolveThemeClass(theme);
-
-  return (
-    <div
-      className={cn(
-        themeClass,
-        'flex min-h-screen w-full items-center justify-center bg-background text-foreground'
-      )}
-      style={{ fontFamily: fontFamily.base }}
-    >
-      <div className="flex flex-col items-center gap-3 text-center">
-        <h1 className="text-2xl font-bold tracking-tight text-foreground">Brand Motion</h1>
-        <p className="text-base text-muted-foreground">Coming soon</p>
-      </div>
-    </div>
-  );
-}
-
-// ============================================================================
-// Story
-// ============================================================================
-
-export const Default: Story = {
-  render: (_, { globals }) => <ComingSoon theme={globals.theme || 'future-dark'} />,
-};
